requirements.txt
google-generativeai
python-dotenv
radon
pytest
colorama
googlesearch-python
gitpython